visualisation/requirements.txt
# BioOptimize Fungi — Streamlit UI + COBRApy + NCBI search
streamlit>=1.29.0
pandas>=1.5.0
cobra>=0.29.0
plotly>=5.18.0
biopython>=1.81
